requirements.txt
pytest==6.2.4
pytest-xdist==2.3.0
docker==5.0.0
ansible==4.2.0
ansible-runner==2.0.1
psycopg2-binary==2.9.1
pytest-html==3.1.1